File reorganizing and adding documentation
