--- TAScheduler/NonDjangoClasses/course_lab_management/CourseManagement.py
@@ -1,27 +1,0 @@
-# createCourse(courseName, courseTime, courseDays, courseHours, courseInstructor, courseTA)
-# editCourse(courseName, courseTime, courseDays, courseHours, courseInstructor, courseTa)
-# deleteCourse(courseName)
-# populateSearchClass(searchPromp)  should change class to course
-# displayAllCourse()
-
-class CourseManagement:
-
-    @staticmethod
-    def createCourse(courseName, courseTime, courseDays, courseHours, courseInstructor, courseTA):
-        pass
-
-    @staticmethod
-    def editCourse(course, courseName, courseTime, courseDays, courseHours, courseInstructor, courseTA):
-        pass
-
-    @staticmethod
-    def deleteCourse(course, courseName):
-        pass
-
-    @staticmethod
-    def populateSearchClass(searchPromp):
-        pass
-
-    @staticmethod
-    def displayAllCourse():
-        pass
