+++ TAScheduler/User/UserType.py
@@ -1,0 +1,15 @@
+class UserTyoe:
+
+	# Preconditions: The user has to have been instantiated
+	# Postconditions: Sets the Type of the user
+	# Side-effects: None
+	# User Type(in) - Type of the use
+    def setType():
+        pass
+
+	# Preconditions: The user has to have been instantiated
+	# Postconditions: Gets the Type of the user
+	# Side-effects: None
+	# User Type(out) - Type of the user
+    def getType():
+        pass
