+++ TAScheduler/tests/LabManagementTests.py
@@ -1,6 +1,6 @@
 import unittest
 
-from TAScheduler.NonDjangoClasses.user_and_login.User import User
+from TAScheduler.User.User import User
 
 
 class CreateLabTests(unittest.TestCase):
